requirements.txt
asgiref==3.5.2
black==22.8.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
pytz==2022.2.1
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.4
urllib3==1.26.12
Werkzeug==2.2.2
whitenoise==6.2.0
